Fetch data to place markers on map.

+++ db/migrate/20180126215152_create_postings.rb
@@ -1,8 +1,8 @@
 class CreatePostings < ActiveRecord::Migration[5.1]
   def change
     create_table :postings do |t|
-      t.decimal :latitude, null: false
-      t.decimal :longitude, null: false
+      t.float :latitude, null: false
+      t.float :longitude, null: false
       t.string :name
       t.text :description
       t.integer :likes
